+++ packages/core/src/bridge.js
@@ -520,6 +520,7 @@ export class Bridge {
     } else {
       $bridge = {
         // bridge
+        getBridgeEventManager: () => this.getBridgeEventManager(),
         logout: () => this.logout(),
         subscribeToEvent: (/** @type {string} */ eventName, /** @type {Function} */ callback) =>
           this.subscribeToEvent(eventName, callback),
@@ -680,6 +681,15 @@ export class Bridge {
    */
   handleBack() {
     this.goBack();
+  }
+
+  /**
+   * Returns the event manager.
+   *
+   * @returns {Object} The event manager.
+   */
+  getBridgeEventManager() {
+    return eventManager;
   }
 
   /**