+++ packages/example/blank-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # blank-app
 
+## 0.1.5-rc.0
+
+### Patch Changes
+
+- Updated dependencies [c17c679]
+  - @open-cells/core@1.2.0-rc.0
+  - @open-cells/element-controller@1.0.5-rc.0
+  - @open-cells/page-controller@1.0.5-rc.0
+
 ## 0.1.4
 
 ### Patch Changes